refactor bonding requests
small recfactors for wallet smoosh

fix various type errors

fix various type errors

more code untangling, yay
