+++ nym-wallet/src/context/bonding.tsx
@@ -21,6 +21,7 @@ import {
   TBondMixnodeSignatureArgs,
   TUpdateBondArgs,
   TNodeDescription,
+  isNymNode,
 } from 'src/types';
 import { Console } from 'src/utils/console';
 import {
@@ -66,6 +67,7 @@ import {
   toPercentIntegerString,
   unymToNym,
 } from '../utils';
+import useGetNodeDetails from 'src/hooks/useGetNodeDetails';
 
 export type TBondedMixnode = {
   name?: string;
@@ -79,13 +81,13 @@ export type TBondedMixnode = {
   operatorRewards?: DecCoin;
   delegators: number;
   status: MixnodeStatus;
-  proxy?: string;
+  proxy?: string | null;
   operatorCost: DecCoin;
   host: string;
   estimatedRewards?: DecCoin;
   activeSetProbability?: SelectionChance;
   standbySetProbability?: SelectionChance;
-  routingScore: number;
+  routingScore?: number;
   httpApiPort: number;
   mixPort: number;
   verlocPort: number;
@@ -96,16 +98,14 @@ export type TBondedMixnode = {
 
 export interface TBondedGateway {
   name?: string;
-  id: number;
   identityKey: string;
   ip: string;
   bond: DecCoin;
-  location?: string;
-  proxy?: string;
+  location: string;
+  proxy: string | null;
   host: string;
   httpApiPort: number;
   mixPort: number;
-  verlocPort: number;
   version: string;
   routingScore?: {
     current: number;
@@ -113,12 +113,18 @@ export interface TBondedGateway {
   };
 }
 
+export type TNymNode = {
+  nodeId: number;
+};
+
+export type TBondedNode = TBondedMixnode | TBondedGateway | TNymNode;
+
 export type TokenPool = 'locked' | 'balance';
 
 export type TBondingContext = {
   isLoading: boolean;
   error?: string;
-  bondedNode?: TBondedMixnode | TBondedGateway;
+  bondedNode?: TBondedMixnode | TBondedGateway | TNymNode | null;
   refresh: () => Promise<void>;
   bondMixnode: (data: TBondMixNodeArgs, tokenPool: TokenPool) => Promise<TransactionExecuteResult | undefined>;
   bondGateway: (data: TBondGatewayArgs, tokenPool: TokenPool) => Promise<TransactionExecuteResult | undefined>;
@@ -168,11 +174,12 @@ export const BondingContext = createContext<TBondingContext>({
 export const BondingContextProvider: FCWithChildren = ({ children }): JSX.Element => {
   const [isLoading, setIsLoading] = useState(false);
   const [error, setError] = useState<string>();
-  const [bondedNode, setBondedNode] = useState<TBondedMixnode | TBondedGateway>();
+
   const [isVestingAccount, setIsVestingAccount] = useState(false);
 
   const { userBalance, clientDetails } = useContext(AppContext);
   const { ownership, isLoading: isOwnershipLoading } = useCheckOwnership();
+  const { bondedNode } = useGetNodeDetails(clientDetails?.client_address || '');
 
   useEffect(() => {
     userBalance.fetchBalance();
@@ -186,266 +193,13 @@ export const BondingContextProvider: FCWithChildren = ({ children }): JSX.Elemen
 
   const resetState = () => {
     setError(undefined);
-    setBondedNode(undefined);
-  };
-
-  /**
-   * Fetch mixnode **optional** data.
-   * ⚠ The underlying queries are allowed to fail.
-   */
-  const fetchMixnodeDetails = async (mixId: number, host: string, port: number) => {
-    const details: {
-      status: MixnodeStatus;
-      stakeSaturation: string;
-      estimatedRewards?: DecCoin;
-      uptime: number;
-      averageUptime?: number;
-      setProbability?: InclusionProbabilityResponse;
-      nodeDescription?: TNodeDescription | undefined;
-      operatorRewards?: DecCoin;
-      uncappedSaturation?: number;
-    } = {
-      status: 'not_found',
-      stakeSaturation: '0',
-      uptime: 0,
-    };
-
-    const statusReq: TauriReq<typeof getMixnodeStatus> = {
-      name: 'getMixnodeStatus',
-      request: () => getMixnodeStatus(mixId),
-      onFulfilled: (value) => {
-        details.status = value.status;
-      },
-    };
-
-    const uptimeReq: TauriReq<typeof getMixnodeUptime> = {
-      name: 'getMixnodeUptime',
-      request: () => getMixnodeUptime(mixId),
-      onFulfilled: (value) => {
-        details.uptime = value;
-      },
-    };
-
-    const stakeSaturationReq: TauriReq<typeof getMixnodeStakeSaturation> = {
-      name: 'getMixnodeStakeSaturation',
-      request: () => getMixnodeStakeSaturation(mixId),
-      onFulfilled: (value) => {
-        details.stakeSaturation = decimalToPercentage(value.saturation);
-        const rawUncappedSaturation = decimalToFloatApproximation(value.uncapped_saturation);
-        if (rawUncappedSaturation && rawUncappedSaturation > 1) {
-          details.uncappedSaturation = Math.round(rawUncappedSaturation * 100);
-        }
-      },
-    };
-
-    const rewardReq: TauriReq<typeof getMixnodeRewardEstimation> = {
-      name: 'getMixnodeRewardEstimation',
-      request: () => getMixnodeRewardEstimation(mixId),
-      onFulfilled: (value) => {
-        const estimatedRewards = unymToNym(value.estimation.total_node_reward);
-        if (estimatedRewards) {
-          details.estimatedRewards = {
-            amount: estimatedRewards,
-            denom: 'nym',
-          };
-        }
-      },
-    };
-
-    const inclusionReq: TauriReq<typeof getInclusionProbability> = {
-      name: 'getInclusionProbability',
-      request: () => getInclusionProbability(mixId),
-      onFulfilled: (value) => {
-        details.setProbability = value;
-      },
-    };
-
-    const avgUptimeReq: TauriReq<typeof getMixnodeAvgUptime> = {
-      name: 'getMixnodeAvgUptime',
-      request: () => getMixnodeAvgUptime(),
-      onFulfilled: (value) => {
-        details.averageUptime = value as number | undefined;
-      },
-    };
-
-    const nodeDescReq: TauriReq<typeof getNodeDescriptionRequest> = {
-      name: 'getNodeDescription',
-      request: () => getNodeDescriptionRequest(host, port),
-      onFulfilled: (value) => {
-        details.nodeDescription = value;
-      },
-    };
-
-    const operatorRewardsReq: TauriReq<typeof getPendingOperatorRewards> = {
-      name: 'getPendingOperatorRewards',
-      request: () => getPendingOperatorRewards(clientDetails?.client_address || ''),
-      onFulfilled: (value) => {
-        details.operatorRewards = decCoinToDisplay(value);
-      },
-    };
-
-    await fireRequests([
-      statusReq,
-      uptimeReq,
-      stakeSaturationReq,
-      rewardReq,
-      inclusionReq,
-      avgUptimeReq,
-      nodeDescReq,
-      operatorRewardsReq,
-    ]);
-
-    return details;
-  };
-
-  /**
-   * Fetch gateway **optional** data.
-   * ⚠ The underlying queries are allowed to fail.
-   */
-  const fetchGatewayDetails = async (identityKey: string, host: string, port: number) => {
-    const details: {
-      routingScore?: { current: number; average: number } | undefined;
-      nodeDescription?: TNodeDescription | undefined;
-    } = {};
-
-    const reportReq: TauriReq<typeof getGatewayReport> = {
-      name: 'getGatewayReport',
-      request: () => getGatewayReport(identityKey),
-      onFulfilled: (value) => {
-        details.routingScore = { current: value.most_recent, average: value.last_day };
-      },
-    };
-
-    const nodeDescReq: TauriReq<typeof getNodeDescriptionRequest> = {
-      name: 'getNodeDescription',
-      request: () => getNodeDescriptionRequest(host, port),
-      onFulfilled: (value) => {
-        details.nodeDescription = value;
-      },
-    };
-
-    await fireRequests([reportReq, nodeDescReq]);
-
-    return details;
-  };
-
-  const calculateStake = (pledge: string, delegations: string) => {
-    let stake;
-    try {
-      stake = unymToNym(Big(pledge).plus(delegations));
-    } catch (e: any) {
-      Console.warn(`not a valid decimal number: ${e}`);
-    }
-    return stake;
-  };
-
-  const refresh = useCallback(async () => {
-    setIsLoading(true);
-    setError(undefined);
-
-    if (ownership.hasOwnership && ownership.nodeType === EnumNodeType.mixnode && clientDetails) {
-      try {
-        const data = await getMixnodeBondDetails();
-        if (data) {
-          const {
-            bond_information,
-            rewarding_details,
-            bond_information: { mix_id },
-          } = data;
-
-          const {
-            status,
-            stakeSaturation,
-            uncappedSaturation: uncappedStakeSaturation,
-            estimatedRewards,
-            uptime,
-            operatorRewards,
-            averageUptime,
-            nodeDescription,
-            setProbability,
-          } = await fetchMixnodeDetails(
-            mix_id,
-            bond_information.mix_node.host,
-            bond_information.mix_node.http_api_port,
-          );
-
-          setBondedNode({
-            id: data.bond_information.mix_id,
-            name: nodeDescription?.name,
-            mixId: mix_id,
-            identityKey: bond_information.mix_node.identity_key,
-            stake: {
-              amount: calculateStake(rewarding_details.operator, rewarding_details.delegates),
-              denom: bond_information.original_pledge.denom,
-            },
-            bond: decCoinToDisplay(bond_information.original_pledge),
-            profitMargin: toPercentIntegerString(rewarding_details.cost_params.profit_margin_percent),
-            delegators: rewarding_details.unique_delegations,
-            proxy: bond_information.proxy,
-            operatorRewards,
-            uptime,
-            status,
-            stakeSaturation,
-            uncappedStakeSaturation,
-            operatorCost: decCoinToDisplay(rewarding_details.cost_params.interval_operating_cost),
-            host: bond_information.mix_node.host.replace(/\s/g, ''),
-            routingScore: averageUptime,
-            activeSetProbability: setProbability?.in_active,
-            standbySetProbability: setProbability?.in_reserve,
-            estimatedRewards,
-            httpApiPort: bond_information.mix_node.http_api_port,
-            mixPort: bond_information.mix_node.mix_port,
-            verlocPort: bond_information.mix_node.verloc_port,
-            version: bond_information.mix_node.version,
-            isUnbonding: bond_information.is_unbonding,
-          } as TBondedMixnode);
-        }
-      } catch (e: any) {
-        Console.warn(e);
-        setError(`While fetching current bond state, an error occurred: ${e}`);
-      }
-    }
-
-    if (ownership.hasOwnership && ownership.nodeType === EnumNodeType.gateway) {
-      try {
-        const data = await getGatewayBondDetails();
-        if (data) {
-          const { gateway, proxy } = data;
-          const { nodeDescription, routingScore } = await fetchGatewayDetails(
-            gateway.identity_key,
-            data.gateway.host,
-            data.gateway.clients_port,
-          );
-          setBondedNode({
-            name: nodeDescription?.name,
-            identityKey: gateway.identity_key,
-            mixPort: gateway.mix_port,
-            httpApiPort: gateway.clients_port,
-            host: gateway.host,
-            ip: gateway.host,
-            location: gateway.location,
-            bond: decCoinToDisplay(data.pledge_amount),
-            proxy,
-            routingScore,
-            version: gateway.version,
-          } as TBondedGateway);
-        }
-      } catch (e: any) {
-        Console.warn(e);
-        setError(`While fetching current bond state, an error occurred: ${e}`);
-      }
-    }
-
-    if (!ownership.hasOwnership) {
-      resetState();
-    }
     setIsLoading(false);
-  }, [ownership]);
+  };
 
-  useEffect(() => {
-    refresh();
-  }, [ownership, refresh]);
-
+  const refresh = () => {
+    resetState();
+    return Promise.resolve();
+  };
   const bondMixnode = async (data: TBondMixNodeArgs, tokenPool: TokenPool) => {
     let tx: TransactionExecuteResult | undefined;
     setIsLoading(true);
@@ -517,7 +271,7 @@ export const BondingContextProvider: FCWithChildren = ({ children }): JSX.Elemen
     try {
       // JS: this check is not entirely valid. you can have proxy field set whilst not using the vesting contract,
       // you have to check if proxy exists AND if it matches the known vesting contract address!
-      if (bondedNode?.proxy) {
+      if (bondedNode && isMixnode(bondedNode) && bondedNode.proxy) {
         tx = await updateMixnodeVestingCostParamsRequest(defaultCostParams, fee?.fee);
       } else {
         tx = await updateMixnodeCostParamsRequest(defaultCostParams, fee?.fee);
@@ -535,7 +289,7 @@ export const BondingContextProvider: FCWithChildren = ({ children }): JSX.Elemen
     let tx;
     setIsLoading(true);
     try {
-      if (bondedNode?.proxy) tx = await vestingClaimOperatorReward(fee?.fee);
+      if (bondedNode && !isNymNode(bondedNode)) tx = await vestingClaimOperatorReward(fee?.fee);
       else tx = await claimOperatorReward(fee?.fee);
     } catch (e: any) {
       setError(`an error occurred: ${e}`);
