+++ nym-wallet/src/requests/actions.ts
@@ -7,14 +7,7 @@ import {
   MixNodeCostParams,
   GatewayConfigUpdate,
 } from '@nymproject/types';
-import {
-  EnumNodeType,
-  TBondGatewayArgs,
-  TBondGatewaySignatureArgs,
-  TBondMixNodeArgs,
-  TBondMixnodeSignatureArgs,
-  TUpdateBondArgs,
-} from '../types';
+import { TBondGatewayArgs, TBondGatewaySignatureArgs } from '../types';
 import { invokeWrapper } from './wrapper';
 
 export const bondGateway = async (args: TBondGatewayArgs) =>
@@ -22,19 +15,6 @@ export const bondGateway = async (args: TBondGatewayArgs) =>
 
 export const generateGatewayMsgPayload = async (args: Omit<TBondGatewaySignatureArgs, 'tokenPool'>) =>
   invokeWrapper<string>('generate_gateway_bonding_msg_payload', args);
-
-export const unbondGateway = async (fee?: Fee) => invokeWrapper<TransactionExecuteResult>('unbond_gateway', { fee });
-
-export const bondMixNode = async (args: TBondMixNodeArgs) =>
-  invokeWrapper<TransactionExecuteResult>('bond_mixnode', args);
-
-export const generateMixnodeMsgPayload = async (args: Omit<TBondMixnodeSignatureArgs, 'tokenPool'>) =>
-  invokeWrapper<string>('generate_mixnode_bonding_msg_payload', args);
-
-export const unbondMixNode = async (fee?: Fee) => invokeWrapper<TransactionExecuteResult>('unbond_mixnode', { fee });
-
-export const updateMixnodeCostParams = async (newCosts: MixNodeCostParams, fee?: Fee) =>
-  invokeWrapper<TransactionExecuteResult>('update_mixnode_cost_params', { newCosts, fee });
 
 export const updateMixnodeConfig = async (update: MixNodeConfigUpdate, fee?: Fee) =>
   invokeWrapper<TransactionExecuteResult>('update_mixnode_config', { update, fee });
@@ -44,13 +24,5 @@ export const updateGatewayConfig = async (update: GatewayConfigUpdate, fee?: Fee
 
 export const send = async (args: { amount: DecCoin; address: string; memo: string; fee?: Fee }) =>
   invokeWrapper<SendTxResult>('send', args);
-
-export const unbond = async (type: EnumNodeType) => {
-  if (type === EnumNodeType.mixnode) return unbondMixNode();
-  return unbondGateway();
-};
-
-export const updateBond = async (args: TUpdateBondArgs) =>
-  invokeWrapper<TransactionExecuteResult>('update_pledge', args);
 
 export const migrateVestedMixnode = async () => invokeWrapper<TransactionExecuteResult>('migrate_vested_mixnode');
