+++ ts-packages/types/src/types/rust/index.ts
@@ -23,7 +23,6 @@ export * from './GatewayConfigUpdate';
 export * from './GatewayCoreStatusResponse';
 export * from './InclusionProbabilityResponse';
 export * from './IntervalRewardingParamsUpdate';
-export * from './IntervalRewardParams';
 export * from './Mixnode';
 export * from './MixNodeBond';
 export * from './MixNodeConfigUpdate';
@@ -57,5 +56,4 @@ export * from './VestingPeriod';
 export * from './WrappedDelegationEvent';
 export * from './NymNode';
 export * from './NymNodeBond';
-export * from './NymNodeDetails';
 export * from './NodeRewarding';
